lmstudio_captioner/requirements.txt

flask==3.0.3
requests==2.32.3
opencv-python==4.10.0.84
pillow==10.4.0
